NPC/Creature update.
Creature and NPC spawning dynamically at last utilize database tables of their own.
No more connections to the “persistent” tables.

NPC's, especially Lairs and Creatures, spawn directly from the database. Scripting is still supported.
Spawns, regardless if it's via the Database or from Script, are now streamlined to use the same core functionality in the code.

Creatures now take full advantage of the attribute concept, making them more accessible to all developers interested in playing with them.

Even though there are still issues, we have some improvements to share with you.

Damage, aggro and XP-system have also been refined, now allowing all npc's to be able to handle combat statistics.
As a direct consequence of the above, players will now get weapon and combat XP when killing lairs.

The re-spawn timer for creatures will now start when the body disappears from the world, not when the creature is killed.

The regeneration rate for creature HAM has been significantly lowered.
To match this, some "faked" combat boosts given to players, are removed.

We changed the type of CDEF Pistol in the Tutorial. The one used now has slightly lower damage stats.

This commit has to be merged with the trunk before you will see it on TC.
May the force be with us.