Fixes an issue with servers always loading.
Justifications:

admin - There is no admin server anymore so this shouldn't be set to active (this entry should be removed altogether in the future).
login - The login server is not galaxy dependent and should not affect a galaxies state.
connection - The connection server at this point in time is the central hub, and doesn't connect to itself thus never increasing the active count.